requirements.txt
accelerate==1.10.1
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
anytree==2.13.0
asttokens==3.0.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bert-score==0.3.13
bioc==2.1
bleach==6.2.0
boto3==1.40.42
botocore==1.40.46
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
conllu==4.5.3
contourpy==1.3.3
cryptography==46.0.1
cwe2==3.0.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.1.1
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
docopt==0.6.2
editdistance==0.8.1
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.1
faiss-cpu==1.12.0
fastjsonschema==2.21.2
filelock==3.19.1
fitz==0.0.1.dev2
flair==0.15.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.12.0
ftfy==6.3.1
gdown==5.2.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
intervaltree==3.1.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json_repair==0.44.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-neo4j==0.5.0
langchain-ollama==0.3.8
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.32
language_tool_python==2.9.4
lazy_loader==0.4
lemminflect==0.2.3
looseversion==1.3.0
lru-dict==1.3.0
lxml==6.0.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mmh3==5.2.0
more-itertools==10.8.0
mpld3==0.5.11
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j==5.28.2
neo4j-graphrag==1.10.0
neontology==2.2.2
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
num2words==0.5.14
numba==0.62.1
numpy==2.2.6
olefile==0.47
ollama==0.5.4
ontolocy==0.9.1
opencv-python==4.12.0.88
OpenHowNet==2.0
openpyxl==3.1.5
opentelemetry-api==1.37.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathlib==1.0.1
pillow==11.3.0
pinyin==0.4.0
pipreq==0.4
platformdirs==4.5.0
pptree==3.1
prompt_toolkit==3.0.52
propcache==0.4.0
protobuf==6.32.1
prov==2.1.1
psutil==7.1.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-oxmsg==0.0.2
pytorch_revgrad==0.2.0
pytz==2025.2
pywin32==311
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.14.1
rdflib==7.2.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
segtok==1.5.11
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sqlitedict==2.1.0
stack-data==0.6.3
stix2==3.0.1
stix2-patterns==2.0.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminaltables==3.1.10
textattack==0.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformer-smaller-training-vocab==0.4.2
transformers==4.57.0
typer==0.19.2
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchfiles==1.1.0
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Wikipedia-API==0.8.1
word2number==1.1
wrapt==1.17.3
xmltodict==1.0.2
xxhash==3.6.0
yarg==0.1.9
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
